Script_Adi
To create the resources with Azure PowerShell, we need to execute a single script that will create a resource group, App Service Plan, and Web App. The script starts with parameters and then executes three cmdlets to create everything we need for deployment. Make sure you are connected to Azure in Azure PowerShell: